Move DAO-based actions into their own namespace
