+++ Civi/Api4/Action/Contact/Create.php
@@ -2,7 +2,7 @@
 
 namespace Civi\Api4\Action\Contact;
 
-use Civi\Api4\Action\Create as DefaultCreate;
+use Civi\Api4\Generic\Action\DAO\Create as DefaultCreate;
 
 /**
  * @inheritDoc
