+++ Civi/Api4/Action/CustomValue/Create.php
@@ -5,7 +5,7 @@ namespace Civi\Api4\Action\CustomValue;
 /**
  * @inheritDoc
  */
-class Create extends \Civi\Api4\Action\Create {
+class Create extends \Civi\Api4\Generic\Action\DAO\Create {
   use \Civi\Api4\Generic\CustomValueCRUD;
 
 }
