+++ Civi/Api4/Action/CustomValue/Delete.php
@@ -5,7 +5,7 @@ namespace Civi\Api4\Action\CustomValue;
 /**
  * Delete one or more items, based on criteria specified in Where param.
  */
-class Delete extends \Civi\Api4\Action\Delete {
+class Delete extends \Civi\Api4\Generic\Action\DAO\Delete {
   use \Civi\Api4\Generic\CustomValueCRUD;
 
 }
