+++ Civi/Api4/Action/CustomValue/Get.php
@@ -5,7 +5,7 @@ namespace Civi\Api4\Action\CustomValue;
 /**
  * Get fields for a custom group.
  */
-class Get extends \Civi\Api4\Action\Get {
+class Get extends \Civi\Api4\Generic\Action\DAO\Get {
   use \Civi\Api4\Generic\CustomValueCRUD;
 
 }
