+++ Civi/Api4/Action/CustomValue/GetFields.php
@@ -9,7 +9,7 @@ use Civi\Api4\Service\Spec\SpecFormatter;
 /**
  * Get fields for a custom group.
  */
-class GetFields extends \Civi\Api4\Action\GetFields {
+class GetFields extends \Civi\Api4\Generic\Action\DAO\GetFields {
   use \Civi\Api4\Generic\CustomValueCRUD;
 
   public function _run(Result $result) {
