+++ Civi/Api4/Action/CustomValue/Replace.php
@@ -5,7 +5,7 @@ namespace Civi\Api4\Action\CustomValue;
 /**
  * Given a set of records, will appropriately update the database.
  */
-class Replace extends \Civi\Api4\Action\Replace {
+class Replace extends \Civi\Api4\Generic\Action\DAO\Replace {
   use \Civi\Api4\Generic\CustomValueCRUD;
 
   protected $select = ['id', 'entity_id'];
