+++ Civi/Api4/Action/GetActions.php
@@ -3,7 +3,7 @@
 namespace Civi\Api4\Action;
 
 use Civi\API\Exception\NotImplementedException;
-use Civi\Api4\Generic\AbstractAction;
+use Civi\Api4\Generic\Action\AbstractAction;
 use Civi\Api4\Generic\Result;
 use Civi\Api4\Utils\ReflectionUtils;
