+++ Civi/Api4/Action/Navigation/Get.php
@@ -2,7 +2,7 @@
 
 namespace Civi\Api4\Action\Navigation;
 
-use Civi\Api4\Action\Get as DefaultGet;
+use Civi\Api4\Generic\Action\DAO\Get as DefaultGet;
 
 /**
  * @inheritDoc
