+++ Civi/Api4/Event/GetSpecEvent.php
@@ -2,7 +2,7 @@
 
 namespace Civi\Api4\Event;
 
-use Civi\Api4\Generic\AbstractAction;
+use Civi\Api4\Generic\Action\AbstractAction;
 use Symfony\Component\EventDispatcher\Event as BaseEvent;
 
 class GetSpecEvent extends BaseEvent {
