+++ Civi/Api4/Provider/ActionObjectProvider.php
@@ -29,7 +29,7 @@ namespace Civi\Api4\Provider;
 
 use Civi\API\Event\ResolveEvent;
 use Civi\API\Provider\ProviderInterface;
-use Civi\Api4\Generic\AbstractAction;
+use Civi\Api4\Generic\Action\AbstractAction;
 use Civi\API\Events;
 use Civi\Api4\Generic\Result;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
