+++ tests/phpunit/Mock/Api4/Action/MockArrayEntity/GetFields.php
@@ -2,7 +2,7 @@
 
 namespace Civi\Api4\Action\MockArrayEntity;
 use Civi\Api4\Generic\Result;
-use \Civi\Api4\Action\GetFields as GenericGetFields;
+use Civi\Api4\Generic\Action\DAO\GetFields as GenericGetFields;
 
 /**
  * Get fields
